remove SongTable intermediate component and make table column rendering more flexible
