+++ projects/frontend/src/pages/share/MergedSongs.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { useMergedSongs } from '../../graphql/queries/merged-songs-query'
 import { SongsView } from './SongsView';
 import { LoadingSpinner } from '../../components/common/LoadingSpinner';
+import { SongTableColumn } from '../../components/song-table/song-table-columns';
 
 export const MergedSongs: React.FC = () => {
 	const { loading, error, data: songs } = useMergedSongs()
@@ -11,5 +12,16 @@ export const MergedSongs: React.FC = () => {
 	}
 	if (error) return <div>`Error!: ${error}`</div>;
 
-	return <SongsView title="All songs" songs={songs} />
+	return (
+		<SongsView
+			title="All songs"
+			songs={songs}
+			columns={[
+				SongTableColumn.Title,
+				SongTableColumn.Time,
+				SongTableColumn.Artists,
+				SongTableColumn.Genres
+			]}
+		/>
+	)
 }
